# backend/openapi.yaml
openapi: 3.1.0
info:
  title: SSI Score API
  description: API for managing Social Selling Index (SSI) scores.
  version: 1.0.0
servers:
  - url: http://localhost:4444
    description: Local development server

components:
  schemas:
    SsiScoreInput:
      type: object
      properties:
        ssi_overall_score:
          type: number
          format: float
          description: Overall SSI score.
        establish_your_professional_brand:
          type: number
          format: float
          description: Score for establishing professional brand.
        find_the_right_people:
          type: number
          format: float
          description: Score for finding the right people.
        engage_with_insights:
          type: number
          format: float
          description: Score for engaging with insights.
        build_relationships:
          type: number
          format: float
          description: Score for building relationships.
      required:
        - ssi_overall_score
        - establish_your_professional_brand
        - find_the_right_people
        - engage_with_insights
        - build_relationships
    SsiScore:
      allOf:
        - $ref: "#/components/schemas/SsiScoreInput"
        - type: object
          properties:
            id:
              type: integer
              description: Unique identifier for the SSI score record.
              readOnly: true
            date_created:
              type: string
              format: date-time
              description: Timestamp when the record was created.
              readOnly: true
    Error:
      type: object
      properties:
        error:
          type: string
          description: A brief description of the error.
        details:
          type: string
          description: Detailed error message (optional).

paths:
  /:
    get:
      summary: Health check
      description: Returns the operational status of the API.
      responses:
        "200":
          description: API is operational.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: OK
  /db-check:
    get:
      summary: Database connection check
      description: Checks the connection to the PostgreSQL database.
      responses:
        "200":
          description: Database connection successful.
          content:
            application/json:
              schema:
                type: object
                properties:
                  dbTime:
                    type: string
                    format: date-time
                    description: Current time from the database server.
        "500":
          description: Database connection failed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples: # Use plural 'examples'
                default: # Named example
                  value:
                    error: "Database connection failed"
                    details: "Database error message"
  /ssi:
    post:
      summary: Create SSI score
      description: Creates a new SSI score record in the database.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SsiScoreInput"
      responses:
        "201":
          description: SSI score created successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SsiScore"
        "400":
          description: Invalid input data.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples: # Use plural 'examples'
                default: # Named example
                  value:
                    error: "All SSI parameters must be numbers."
        "500":
          description: Failed to save SSI score.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples: # Use plural 'examples'
                default: # Named example
                  value:
                    error: "Failed to save SSI score"
                    details: "Database error message"
    get:
      summary: Get all SSI scores
      description: Retrieves all SSI score records from the database, ordered by creation date descending.
      responses:
        "200":
          description: A list of SSI scores.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SsiScore"
        "500":
          description: Failed to fetch SSI scores.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples: # Use plural 'examples'
                default: # Named example
                  value:
                    error: "Failed to fetch SSI scores"
                    details: "Database error message"